Create workers.js
This code listens for an HTTP request using the fetch event, and uses the handleRequest function to process the request. The function uses the URL API to extract the IPFS hash from the query parameters of the URL. It then makes an HTTP request to the web3.storages gateway to retrieve the IPFS object, and returns the response. If there is an error, the function returns an error response.